+++ src/types/ProjectDataReference.ts
@@ -2,5 +2,5 @@ export interface ProjectDataReference {
     name: string,
     type: string,
     id: string,
-    last_updated: number
+    lastUpdated: number
 }